requirements.txt
asgiref==3.7.2
Django==5.0.3
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==24.1
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
